Added reference measurements and UI options
At the moment the reference measurement is applied by converting the
complex data to amplitude data and subtracting the amplitude data of the
reference measurement. Phase data is discarded. Reference measurements
are added by adding a file named "ref.txt" in the measurement directory
containing a path to the reference measurement.

Also added options to Huge Plot for coloring measurements by class and
hiding the legend by setting the optional arguments 'classColors' and
'showLegend' to true or false.